Refactoring command execution via class structure
